Update regression test data for tests dealing with CONSUME
The CONSUME package's output changes from python 2.7 to 3.5.  For example,
in one test that uses consume to compute consumption, I get the following:

    (Pdb) pp expected['fire_information'][i]['consumption']
    {'summary': {'flaming': 1856.486619372032,
                 'residual': 55.19598745600178,
                 'smoldering': 242.20430971763648,
                 'total': 2153.8869165456695}}
    (Pdb) pp actual['fire_information'][i]['consumption']
    {'summary': {'flaming': 1863.159939372032,
                 'residual': 55.19598745600178,
                 'smoldering': 242.94578971763644,
                 'total': 2161.3017165456704}}

where 'expected' are the results from running in 2.7 and 'actual' are from 3.5.
In another test that uses CONSUME to compute emissions, I get the following:

    (Pdb) pp expected['fire_information'][i]['emissions']
    {'summary': {'PM25': 148.31906224347662, 'total': 148.31906224347662}}
    (Pdb) pp actual['fire_information'][i]['emissions']
    {'summary': {'PM25': 148.84959118347666, 'total': 148.84959118347666}}

I suspect that the differences arrise from the used of '/', which
is always treated as float divistion in 3.5 but is int division in 2.7
(i.e. rounding the results) when ints are included in the equation.
Given that the package was most recently updated for python 3.*, it's probably
safe to assume that the '/' is intended as float division, so that the 3.5
results are in fact more 'correct'.