Adjust hysplit dispersion window based on met availability
Set num_hours so that hysplit is run up through the last hour of continuous met data.
This affects cases where there is a gap in the met data (e.g. dispersion requested
for days 1,2,3, but there are met data for only days 1 and 3) or truncated met
(e.g. dispersion requested or days 1,2,3, but there are met data for only day 1)

If there are no met data for the first hour in the dispersion window, an
excpetion is raised, even if there are met data for later in the dispersion window.