Use max number of tranche processes if parinit file is input or pardump file is written
And use dummy fires for any processes without real fires.
This is done so as to guarantee that the number of processes is the same
in the run that generates the parinit file(s) and the run that reads them
(assuming they use the same configuration), which is required for
proper functioning.

This commit is adapted from Robert's commit in BSF.

Commit also includes some refactoring to use member attrs to simplify
method signatures.